+++ test/pagy/extras/searchkick_test.rb
@@ -3,6 +3,7 @@
 
 require_relative '../../test_helper'
 require_relative '../../test_helper/searchkick'
+require 'pagy/extras/overflow'
 
 SingleCov.covered! unless ENV['SKIP_SINGLECOV']
 
@@ -48,7 +49,8 @@ describe Pagy::Backend do
       pagy.items.must_equal Pagy::VARS[:items]
       pagy.page.must_equal backend.params[:page]
       results.count.must_equal Pagy::VARS[:items]
-      results.must_equal ["R-B-a-41", "R-B-a-42", "R-B-a-43", "R-B-a-44", "R-B-a-45", "R-B-a-46", "R-B-a-47", "R-B-a-48", "R-B-a-49", "R-B-a-50", "R-B-a-51", "R-B-a-52", "R-B-a-53", "R-B-a-54", "R-B-a-55", "R-B-a-56", "R-B-a-57", "R-B-a-58", "R-B-a-59", "R-B-a-60"]    end
+      results.must_equal ["R-B-a-41", "R-B-a-42", "R-B-a-43", "R-B-a-44", "R-B-a-45", "R-B-a-46", "R-B-a-47", "R-B-a-48", "R-B-a-49", "R-B-a-50", "R-B-a-51", "R-B-a-52", "R-B-a-53", "R-B-a-54", "R-B-a-55", "R-B-a-56", "R-B-a-57", "R-B-a-58", "R-B-a-59", "R-B-a-60"]
+    end
 
     it 'paginates results with defaults' do
       pagy, results = backend.send(:pagy_searchkick, SearchkickModel.pagy_search('a').results)
@@ -69,6 +71,17 @@ describe Pagy::Backend do
       pagy.vars[:link_extra].must_equal 'X'
       results.count.must_equal 10
       results.must_equal ["R-b-11", "R-b-12", "R-b-13", "R-b-14", "R-b-15", "R-b-16", "R-b-17", "R-b-18", "R-b-19", "R-b-20"]
+    end
+
+    it 'paginates with overflow' do
+      pagy, results = backend.send(:pagy_searchkick, SearchkickModel.pagy_search('b').results, page: 200, items: 10, link_extra: 'X', overflow: :last_page)
+      pagy.must_be_instance_of Pagy
+      pagy.count.must_equal 1000
+      pagy.items.must_equal 10
+      pagy.page.must_equal 100
+      pagy.vars[:link_extra].must_equal 'X'
+      results.count.must_equal 10
+      results.must_equal ["R-b-991", "R-b-992", "R-b-993", "R-b-994", "R-b-995", "R-b-996", "R-b-997", "R-b-998", "R-b-999", "R-b-1000"]
     end
 
   end